KAD-767: move `[Entity]QuerySortBy` & `[Entity]QuerySortParameter` into separate classes
