+++ history/kadai-simplehistory-rest-spring/src/main/java/io/kadai/simplehistory/rest/TaskHistoryEventApi.java
@@ -2,7 +2,6 @@ package io.kadai.simplehistory.rest;
 
 import io.kadai.common.rest.QueryPagingParameter;
 import io.kadai.simplehistory.impl.task.TaskHistoryQuery;
-import io.kadai.simplehistory.rest.TaskHistoryEventController.TaskHistoryQuerySortParameter;
 import io.kadai.simplehistory.rest.models.TaskHistoryEventPagedRepresentationModel;
 import io.kadai.simplehistory.rest.models.TaskHistoryEventRepresentationModel;
 import io.kadai.spi.history.api.events.task.TaskHistoryEvent;
