+++ history/kadai-simplehistory-rest-spring/src/main/java/io/kadai/simplehistory/rest/TaskHistoryEventController.java
@@ -18,11 +18,8 @@
 
 package io.kadai.simplehistory.rest;
 
-import io.kadai.common.api.BaseQuery.SortDirection;
 import io.kadai.common.api.KadaiEngine;
-import io.kadai.common.api.exceptions.InvalidArgumentException;
 import io.kadai.common.rest.QueryPagingParameter;
-import io.kadai.common.rest.QuerySortBy;
 import io.kadai.common.rest.QuerySortParameter;
 import io.kadai.common.rest.util.QueryParamsValidator;
 import io.kadai.simplehistory.impl.SimpleHistoryServiceImpl;
@@ -30,13 +27,10 @@ import io.kadai.simplehistory.impl.task.TaskHistoryQuery;
 import io.kadai.simplehistory.rest.assembler.TaskHistoryEventRepresentationModelAssembler;
 import io.kadai.simplehistory.rest.models.TaskHistoryEventPagedRepresentationModel;
 import io.kadai.simplehistory.rest.models.TaskHistoryEventRepresentationModel;
-import io.kadai.spi.history.api.events.task.TaskHistoryCustomField;
 import io.kadai.spi.history.api.events.task.TaskHistoryEvent;
 import io.kadai.spi.history.api.exceptions.KadaiHistoryEventNotFoundException;
 import jakarta.servlet.http.HttpServletRequest;
-import java.beans.ConstructorProperties;
 import java.util.List;
-import java.util.function.BiConsumer;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.MediaTypes;
@@ -101,54 +95,5 @@ public class TaskHistoryEventController implements TaskHistoryEventApi {
 
     TaskHistoryEventRepresentationModel taskEventResource = assembler.toModel(resultEvent);
     return new ResponseEntity<>(taskEventResource, HttpStatus.OK);
-  }
-
-  public enum TaskHistoryQuerySortBy implements QuerySortBy<TaskHistoryQuery> {
-    TASK_HISTORY_EVENT_ID(TaskHistoryQuery::orderByTaskHistoryEventId),
-    BUSINESS_PROCESS_ID(TaskHistoryQuery::orderByBusinessProcessId),
-    PARENT_BUSINESS_PROCESS_ID(TaskHistoryQuery::orderByParentBusinessProcessId),
-    TASK_ID(TaskHistoryQuery::orderByTaskId),
-    EVENT_TYPE(TaskHistoryQuery::orderByEventType),
-    CREATED(TaskHistoryQuery::orderByCreated),
-    USER_ID(TaskHistoryQuery::orderByUserId),
-    DOMAIN(TaskHistoryQuery::orderByDomain),
-    WORKBASKET_KEY(TaskHistoryQuery::orderByWorkbasketKey),
-    POR_COMPANY(TaskHistoryQuery::orderByPorCompany),
-    POR_SYSTEM(TaskHistoryQuery::orderByPorSystem),
-    POR_INSTANCE(TaskHistoryQuery::orderByPorInstance),
-    POR_TYPE(TaskHistoryQuery::orderByPorType),
-    POR_VALUE(TaskHistoryQuery::orderByPorValue),
-    TASK_CLASSIFICATION_KEY(TaskHistoryQuery::orderByTaskClassificationKey),
-    TASK_CLASSIFICATION_CATEGORY(TaskHistoryQuery::orderByTaskClassificationCategory),
-    ATTACHMENT_CLASSIFICATION_KEY(TaskHistoryQuery::orderByAttachmentClassificationKey),
-    CUSTOM_1((query, sort) -> query.orderByCustomAttribute(TaskHistoryCustomField.CUSTOM_1, sort)),
-    CUSTOM_2((query, sort) -> query.orderByCustomAttribute(TaskHistoryCustomField.CUSTOM_2, sort)),
-    CUSTOM_3((query, sort) -> query.orderByCustomAttribute(TaskHistoryCustomField.CUSTOM_3, sort)),
-    CUSTOM_4((query, sort) -> query.orderByCustomAttribute(TaskHistoryCustomField.CUSTOM_4, sort)),
-    OLD_VALUE(TaskHistoryQuery::orderByOldValue);
-
-    private final BiConsumer<TaskHistoryQuery, SortDirection> consumer;
-
-    TaskHistoryQuerySortBy(BiConsumer<TaskHistoryQuery, SortDirection> consumer) {
-      this.consumer = consumer;
-    }
-
-    @Override
-    public void applySortByForQuery(TaskHistoryQuery query, SortDirection sortDirection) {
-      consumer.accept(query, sortDirection);
-    }
-  }
-
-  // Unfortunately this class is necessary, since spring can not inject the generic 'sort-by'
-  // parameter from the super class.
-  public static class TaskHistoryQuerySortParameter
-      extends QuerySortParameter<TaskHistoryQuery, TaskHistoryQuerySortBy> {
-
-    @ConstructorProperties({"sort-by", "order"})
-    public TaskHistoryQuerySortParameter(
-        List<TaskHistoryQuerySortBy> sortBy, List<SortDirection> order)
-        throws InvalidArgumentException {
-      super(sortBy, order);
-    }
   }
 }
