+++ rest/kadai-rest-spring/src/main/java/io/kadai/user/rest/UserController.java
@@ -18,12 +18,10 @@
 
 package io.kadai.user.rest;
 
-import io.kadai.common.api.BaseQuery.SortDirection;
 import io.kadai.common.api.exceptions.InvalidArgumentException;
 import io.kadai.common.api.exceptions.NotAuthorizedException;
 import io.kadai.common.api.security.CurrentUserContext;
 import io.kadai.common.rest.QueryPagingParameter;
-import io.kadai.common.rest.QuerySortBy;
 import io.kadai.common.rest.QuerySortParameter;
 import io.kadai.common.rest.RestEndpoints;
 import io.kadai.common.rest.util.QueryParamsValidator;
@@ -38,9 +36,7 @@ import io.kadai.user.rest.assembler.UserSummaryRepresentationModelAssembler;
 import io.kadai.user.rest.models.UserRepresentationModel;
 import io.kadai.user.rest.models.UserSummaryPagedRepresentationModel;
 import jakarta.servlet.http.HttpServletRequest;
-import java.beans.ConstructorProperties;
 import java.util.List;
-import java.util.function.BiConsumer;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.config.EnableHypermediaSupport;
@@ -151,37 +147,5 @@ public class UserController implements UserApi {
     userService.deleteUser(userId);
 
     return ResponseEntity.noContent().build();
-  }
-
-  public enum UserQuerySortBy implements QuerySortBy<UserQuery> {
-    FIRST_NAME(UserQuery::orderByFirstName),
-    LAST_NAME(UserQuery::orderByLastName),
-    ORG_LEVEL_1(UserQuery::orderByOrgLevel1),
-    ORG_LEVEL_2(UserQuery::orderByOrgLevel2),
-    ORG_LEVEL_3(UserQuery::orderByOrgLevel3),
-    ORG_LEVEL_4(UserQuery::orderByOrgLevel4);
-
-    private final BiConsumer<UserQuery, SortDirection> consumer;
-
-    UserQuerySortBy(BiConsumer<UserQuery, SortDirection> consumer) {
-      this.consumer = consumer;
-    }
-
-    @Override
-    public void applySortByForQuery(UserQuery query, SortDirection sortDirection) {
-      consumer.accept(query, sortDirection);
-    }
-  }
-
-  // Unfortunately this class is necessary, since spring can not inject the generic 'sort-by'
-  // parameter from the super class.
-  public static class UserQuerySortParameter
-      extends QuerySortParameter<UserQuery, UserQuerySortBy> {
-
-    @ConstructorProperties({"sort-by", "order"})
-    public UserQuerySortParameter(List<UserQuerySortBy> sortBy, List<SortDirection> order)
-        throws InvalidArgumentException {
-      super(sortBy, order);
-    }
   }
 }
